Fix failures to list self-referenced symlink
Previous fern.vim can't list file entries containing even one
self-referenced symlink because a `find` command invoked in the
background emits an error such as `Too many levels of symbolic links` by
the `-follow` option. The option tries to dereference symlinks in
listing files. But we only need to dereference symlinks on the command
line, but not dereference symlinks in listing files. So we should use
the `-H` option, which doesn't dereference symlinks except on the
command line, instead of the `-follow` option.

This problem happens on Linux, but not macOS. The `find` command
installed in macOS by default doesn't emit errors in the situation
above.