Correct hover popup position on right side drawer
When navigating through nodes in a drawer window, node names that extend
beyond the width of the fern window are shown in a popup to display the
full name to the user.

A regression introduced in 10dd4bb (Improve hover popup behavior, 2022-03-16)
prevented the popup from being positioned correctly in drawer windows
positioned on the right side. As a result, the hover popup behavior was
disabled for right hand drawers in ab237a1 (Merge pull request #419 from
lambdalisue/fix-popup-right, 2022-04-15).

This revision reintroduces this functionality, ensuring that the popup
windows are positioned correctly in all cases.