introduce separate syntax group for leader symbol
Prior to this revision, the leading symbols of a branch or leaf node
inherited the highlight group from the branch or leaf symbol, which
doesn't look appealing when they are assigned different highlight
groups.

This revision introduces a new syntax group FernLeaderSymbol which may
be highlighted separately from the node symbols. Documentation is
updated accordingly.