feat: add support for rendering text with props
Introduce the ability to render text with attached text properties.
Renderers may now also return list of dictionaries where each line of
text can include text properties for additional text styling
capabilities. This change is backwards compatible with existing
renderers; they will continue to work as before. These features,
however, are not supported in Neovim.

The semantics of this feature was inspired by popup_create(), which can
accept a list of dictionaries with a 'text' and 'props' key.