feat: split up hide_cursor option (#491)
This revision introduces changes to the 'g:fern#hide_cursor' option.

Before these changes, this option forcibly disables 'cursorline' when
focus is not on the fern window and unsets 't_ve' to hide the cursor for
the fern window.

In some configurations, it's desirable to hide the cursor but not change
the cursorline configuration. For instance, the user may want to see
what item is selected in the fern window when working in other buffers.

Although it's relatively trivial to skip 'hide_cursor' altogether and
manually update 't_ve', I believe there's value in supporting this in
Fern.

This revision introduces a new option, 'g:fern#hide_cursorline', which
enables/disables the 'cursorline' option on WinEnter/WinLeave, similar
to what 'hide_cursor' did before.

Documentation has been updated to reflect this change.